requirements.txt
pexpect==4.6.0
ptyprocess==0.6.0
PyYAML==3.13
